API/requirements.txt
Django==2.1.4
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
djangorestframework==3.9.0
psycopg2>=2.7,<3.0
virtualenv==16.1.0
virtualenvwrapper-win==1.2.5
requests==2.25.1
